+++ ec2_alb.tf
@@ -3,6 +3,8 @@ resource "aws_lb_target_group" "concourse" {
   port     = 8080
   protocol = "HTTP"
   vpc_id   = var.vpc_id
+
+  tags = merge({ Name = "${var.prefix}-concourse-backend" }, var.alb_tags)
 }
 
 resource "aws_lb_target_group_attachment" "concourse" {
@@ -24,6 +26,8 @@ resource "aws_lb" "concourse" {
   }
 
   enable_deletion_protection = false
+
+  tags = merge({ Name = "${var.prefix}-concourse-alb" }, var.alb_tags)
 }
 
 resource "aws_lb_listener" "concource_http" {